Move log level initialization to log
